db config with env and swagger setup
