+++ .env
@@ -1,14 +1,21 @@
 BCRYPT_SALT=10
-EMAIL_FROM=srinivas@northgaze.com
-HOST=127.0.0.1
-DATABASE_NAME=client_db
-DBPORT=27017
+
+DB_HOST=127.0.0.1
+DB_NAME=aurora
+DB_PORT=27017
+DB_PREFIX=mongodb
+DB_OPTIONS=authSource=admin&authMechanism=DEFAULT
+DB_USERNAME=admin
+DB_PASSWORD=123456
+
 SERVERPORT=8080
 CLIENT_URL=http://localhost:3000
+SERVER_URL=http://localhost:8080/
+
 BASE_AUTH_USERNAME=ClientApi
 BASE_AUTH_PASSWORD=ClientApi
-SERVER_URL=http://localhost:8080/
 NODE_ENV=development
+
 STRIPE_SECRET_KEY='sk_test_w9FFlXYuNXBQAGbAMcbTfvUc'
 STRIPE_PUBLIC_KEY='pk_test_kLXZAIgM9jW9flHREsbzTupH'
 STRIPE_WEBHOOK_SECRET='whsec_7057daa4eb33db727887280e7c0f93917b387893812976790eb65ae7329cf338'
@@ -16,8 +23,5 @@ TWILIO_ACCOUNT_SID='AC165edf7f114aacef138978fa52a1f539'
 TWILIO_AUTH_TOKEN='c0e67d325930e935703b56620e446af5'
 TWILIO_PHONE_NUMBER='+19294941821'
 SENDGRID_API_KEY='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'
-MONGODB_PREFIX=mongodb
-MONGODB_OPTIONS=replicaSet=rs0&maxPoolSize=10&maxIdleTimeMS=60000
-MONGODB_USERNAME=northgaze
-MONGODB_PASSWORD=123456
+EMAIL_FROM=srinivas@northgaze.com
